% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.getCache}
\alias{.getCache}
\title{Get cache}
\usage{
.getCache()
}
\value{
A BiocFileCache object.
}
\description{
\code{.getCache} creates and loads a local cache for the `ete3r`
package.
}
\examples{

\dontrun{

cache <- .getCache()
cache

}


}
