% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{getTree}
\alias{getTree}
\title{Get NCBI tree}
\usage{
getTree(taxids)
}
\arguments{
\item{taxids}{Integer vector of NCBI IDs (taxids).}
}
\value{
A phylogenetic tree (tidytree class object).
}
\description{
\code{getTree} returns an annotated phylogenetic tree from
NCBI IDs (taxids) based on the NCBI taxonomy.
}
\details{
This is a wraper around the ete3 package
\url{http://etetoolkit.org/docs/latest/tutorial/tutorial_ncbitaxonomy.html}.
}
\examples{

\dontrun{

taxids <- c(9606, 9598, 10090, 7707, 8782)
tree <- getTree(taxids)
class(tree)

}

}
